move confetti to after all actions have run
