+++ app/web/src/store/actions.store.ts
@@ -1,6 +1,7 @@
 import { defineStore } from "pinia";
 import * as _ from "lodash-es";
 import { addStoreHooks, ApiRequest } from "@si/vue-lib/pinia";
+import JSConfetti from "js-confetti";
 import { Resource } from "@/api/sdf/dal/resource";
 import { useWorkspacesStore } from "@/store/workspaces.store";
 import { DefaultMap } from "@/utils/defaultmap";
@@ -120,6 +121,16 @@ export type ActionId = string;
 export type FuncRunId = string;
 
 // END STUFF
+
+let jsConfetti: JSConfetti;
+const confettis = [
+  { emojis: ["🎉"] },
+  { emojis: ["🍿"] },
+  { emojis: ["🤘", "🤘🏻", "🤘🏼", "🤘🏽", "🤘🏾", "🤘🏿"] },
+  { emojis: ["❤️", "🧡", "💛", "💚", "💙", "💜"] },
+  { emojis: ["🍾", "🍷", "🍸", "🍹", "🍺", "🥂", "🍻"] },
+  { emojis: ["🏳️‍🌈", "🏳️‍⚧️", "⚡️", "🌈", "✨", "🔥", "🇧🇷"] },
+];
 
 export const useActionsStore = () => {
   const workspacesStore = useWorkspacesStore();
@@ -291,6 +302,9 @@ export const useActionsStore = () => {
                 visibility_change_set_pk: changeSetId,
               },
               onSuccess: (response) => {
+                if (this.actions.length > 0 && response.length === 0)
+                  jsConfetti.addConfetti(_.sample(confettis));
+
                 this.actions = response;
               },
             });
@@ -366,6 +380,11 @@ export const useActionsStore = () => {
           },
         },
         onActivated() {
+          jsConfetti = new JSConfetti({
+            canvas:
+              (document.getElementById("confetti") as HTMLCanvasElement) ||
+              undefined,
+          });
           if (!changeSetId) return;
 
           this.LOAD_ACTIONS();
